[TOA-38] Remove unnecessary interface Add todo for NavigationRequest class
