--- app/src/main/java/pl/droidsonroids/toast/app/NavigatorImpl.kt
@@ -1,18 +1,0 @@
-package pl.droidsonroids.toast.app
-
-import android.content.Context
-import pl.droidsonroids.toast.app.speakers.SpeakersSearchActivity
-import pl.droidsonroids.toast.utils.NavigationRequest
-
-class NavigatorImpl : Navigator {
-    override fun dispatch(context: Context, navigationRequest: NavigationRequest) {
-        when (navigationRequest) {
-            is NavigationRequest.SpeakersSearch -> showSpeakersSearch(context)
-        }
-    }
-
-    private fun showSpeakersSearch(context: Context) {
-        val intent = SpeakersSearchActivity.createIntent(context)
-        context.startActivity(intent)
-    }
-}
